[COMMIT] 03/14 commit 검색기능(개발중)
