+++ src/main/java/com/simulation/LoLItemSimulation/controller/CommentController.java
@@ -8,6 +8,8 @@ import com.simulation.LoLItemSimulation.repository.CommentLikeRepository;
 import com.simulation.LoLItemSimulation.repository.CommentRepository;
 import com.simulation.LoLItemSimulation.service.CommentService;
 import jakarta.servlet.http.HttpServletRequest;
+import lombok.Getter;
+import lombok.Setter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -18,7 +20,7 @@ import java.util.Map;
 
 // CommentController.java
 @Controller
-//@RequestMapping("/comment")
+@RequestMapping("/comment")
 public class CommentController {
   @Autowired
   private CommentService commentService;
@@ -124,9 +126,71 @@ public class CommentController {
     }
   }
 
+  // 댓글 좋아요 갯수
   @GetMapping("/{commentId}/like/count")
   public ResponseEntity<Long> getLikeCount(@PathVariable Long commentId) {
     long likeCount = commentLikeRepository.countLikesByCommentId(commentId);
     return ResponseEntity.ok(likeCount);
   }
+
+  // 댓글 생성
+  @PostMapping(value = "/createComment", consumes = "application/json")
+  public ResponseEntity<Comment> createComment(@RequestBody CommentDto commentDto, HttpServletRequest request) {
+    // 댓글 생성 로직...
+//    log.info("v파라미터 확인 :  " + commentDto.toString());
+    Comment comment = new Comment();
+    comment.setPostId(commentDto.getPostId());
+    comment.setNickname(commentDto.getNickname());
+    comment.setPassword(commentDto.getPassword());
+    comment.setContent(commentDto.getContent());
+
+    // 클라이언트의 실제 IP 주소 가져오기
+    String ipAddress = getClientIP(request);
+    comment.setIpAddress(ipAddress);
+//    log.info("댓글생성 확인 : " + comment);
+    // 댓글을 저장
+    commentRepository.save(comment);
+
+    return ResponseEntity.ok(comment);
+  }
+
+  // 댓글 삭제시 비밀번호 확인
+  @PostMapping("/checkCommentPassword")
+  @ResponseBody
+  public String checkCommentPassword(@RequestBody CommentPasswordRequest request) {
+    Long commentId = request.getCommentId(); // 댓글 ID
+    String password = request.getPassword(); // 입력된 비밀번호
+
+    // commentId와 password를 이용하여 댓글의 비밀번호를 확인하는 로직 작성
+    boolean passwordMatch = commentService.checkCommentPassword(commentId, password);
+
+    return String.valueOf(passwordMatch);
+  }
+  @Getter
+  @Setter
+  static class CommentPasswordRequest {
+    private Long commentId;
+    private String password;
+  }
+
+  // 댓글 삭제
+  @DeleteMapping("/deleteComment/{commentId}")
+  @ResponseBody
+  public String deleteComment(@PathVariable Long commentId) {
+    boolean deleted = commentService.deleteComment(commentId);
+    if (deleted) {
+      return "댓글이 성공적으로 삭제되었습니다.";
+    } else {
+      return "댓글을 삭제하는 중 오류가 발생했습니다.";
+    }
+  }
+
+  // 게시글의 총 댓글 수를 반환하는 엔드포인트
+  @GetMapping("/{postId}/commentCount")
+  public ResponseEntity<Integer> getCommentCount(@PathVariable Long postId) {
+    int commentCount = commentService.countCommentsByPostId(postId);
+    return ResponseEntity.ok(commentCount);
+  }
+
+
 }
